feat: add module, classes, refactor
