+++ index.html
@@ -8,6 +8,6 @@
     <title>codejam-virtual-keyboard</title>
 </head>
 <body>
- <script src="./main.js"></script>   
+ <script src="./main.js" type="module"></script>   
 </body>
 </html>
